fixes #269551
