+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/executeStrategy/richExecuteStrategy.ts
@@ -6,13 +6,14 @@
 import type { CancellationToken } from '../../../../../../base/common/cancellation.js';
 import { CancellationError } from '../../../../../../base/common/errors.js';
 import { Emitter, Event } from '../../../../../../base/common/event.js';
-import { DisposableStore, MutableDisposable, toDisposable, type IDisposable } from '../../../../../../base/common/lifecycle.js';
+import { DisposableStore, MutableDisposable } from '../../../../../../base/common/lifecycle.js';
 import { isNumber } from '../../../../../../base/common/types.js';
 import type { ICommandDetectionCapability, ITerminalCommand } from '../../../../../../platform/terminal/common/capabilities/capabilities.js';
 import { ITerminalLogService } from '../../../../../../platform/terminal/common/terminal.js';
 import type { ITerminalInstance } from '../../../../terminal/browser/terminal.js';
 import { trackIdleOnPrompt, type ITerminalExecuteStrategy, type ITerminalExecuteStrategyResult } from './executeStrategy.js';
 import type { IMarker as IXtermMarker } from '@xterm/xterm';
+import { setupRecreatingStartMarker } from './strategyHelpers.js';
 
 /**
  * This strategy is used when the terminal has rich shell integration/command detection is
@@ -58,31 +59,14 @@ export class RichExecuteStrategy implements ITerminalExecuteStrategy {
 				}),
 			]);
 
-			// Record where the command started. If the marker gets disposed, re-created it where
-			// the cursor is. This can happen in prompts where they clear the line and rerender it
-			// like powerlevel10k's transient prompt
-			const markerListener = new MutableDisposable<IDisposable>();
-			const recreateStartMarker = () => {
-				if (store.isDisposed) {
-					return;
-				}
-				const marker = xterm.raw.registerMarker();
-				this._startMarker.value = marker ?? undefined;
-				this._onDidCreateStartMarker.fire(marker);
-				if (!marker) {
-					markerListener.clear();
-					return;
-				}
-				markerListener.value = marker.onDispose(() => {
-					recreateStartMarker();
-				});
-			};
-			recreateStartMarker();
-			store.add(toDisposable(() => {
-				markerListener.dispose();
-				this._startMarker.clear();
-				this._onDidCreateStartMarker.fire(undefined);
-			}));
+			store.add(this._startMarker);
+			setupRecreatingStartMarker(
+				xterm,
+				this._startMarker,
+				m => this._onDidCreateStartMarker.fire(m),
+				store,
+				this._log.bind(this)
+			);
 
 			// Execute the command
 			this._log(`Executing command line \`${commandLine}\``);
